requirements.txt
Flask
Flask_SQLAlchemy
psycopg2-binary
Flask-Migrate
Flask-Cors
python-dotenv
